Day7 task done

+++ Day5/Q3.cpp
@@ -9,27 +9,27 @@
 #include <map>
 #define ll long long int
 using namespace std;
-vector<vector<int>> fourSum1(vector<int> &nums, int target)
+vector<vector<ll>> fourSum1(vector<ll> &nums, ll target)
 {
-    set<vector<int>> ans;
-    int n = nums.size();
-    for (int i = 0; i < n; i++)
+    set<vector<ll>> ans;
+    ll n = nums.size();
+    for (ll i = 0; i < n; i++)
     {
-        for (int j = i + 1; j < n; j++)
+        for (ll j = i + 1; j < n; j++)
         {
-            for (int k = j + 1; k < n; k++)
+            for (ll k = j + 1; k < n; k++)
             {
-                for (int l = k + 1; l < n; l++)
+                for (ll l = k + 1; l < n; l++)
                 {
                     if (j != i && k != j && l != k)
                     {
-                        long long int sum = nums[i];
+                        ll sum = nums[i];
                         sum += nums[j];
                         sum += nums[k];
                         sum += nums[l];
                         if (sum == target)
                         {
-                            vector<int> temp{nums[i], nums[j], nums[k], nums[l]};
+                            vector<ll> temp{nums[i], nums[j], nums[k], nums[l]};
                             sort(temp.begin(), temp.end());
                             ans.insert(temp);
                         }
@@ -38,7 +38,7 @@ vector<vector<int>> fourSum1(vector<int> &nums, int target)
             }
         }
     }
-    vector<vector<int>> vec_ans;
+    vector<vector<ll>> vec_ans;
     for (auto val : ans)
     {
         vec_ans.push_back(val);
@@ -46,32 +46,32 @@ vector<vector<int>> fourSum1(vector<int> &nums, int target)
     return vec_ans;
 }
 
-vector<vector<int>> fourSum2(vector<int> &nums, int target)
+vector<vector<ll>> fourSum2(vector<ll> &nums, ll target)
 {
-    set<vector<int>> ans;
-    int n = nums.size();
-    unordered_map<int, int> mp;
-    for (int i = 0; i < n; i++)
+    set<vector<ll>> ans;
+    ll n = nums.size();
+    unordered_map<ll, ll> mp;
+    for (ll i = 0; i < n; i++)
     {
         mp[nums[i]] = i;
     }
-    for (int i = 0; i < n; i++)
+    for (ll i = 0; i < n; i++)
     {
-        for (int j = i + 1; j < n; j++)
+        for (ll j = i + 1; j < n; j++)
         {
-            for (int k = j + 1; k < n; k++)
+            for (ll k = j + 1; k < n; k++)
             {
-                long long int sum = nums[i];
+                ll sum = nums[i];
                 sum += nums[j];
                 sum += nums[k];
-                long long int r = target - sum;
+                ll r = target - sum;
                 if (mp.find(r) != mp.end())
                 {
                     auto ptr = mp.find(r);
                     cout << "founded val:" << ptr->first << endl;
                     if (ptr->second != i && ptr->second != j && ptr->second != k)
                     {
-                        vector<int> temp{nums[i], nums[j], nums[k], nums[ptr->second]};
+                        vector<ll> temp{nums[i], nums[j], nums[k], nums[ptr->second]};
                         sort(temp.begin(), temp.end());
                         ans.insert(temp);
                     }
@@ -79,7 +79,7 @@ vector<vector<int>> fourSum2(vector<int> &nums, int target)
             }
         }
     }
-    vector<vector<int>> vec_ans;
+    vector<vector<ll>> vec_ans;
     for (auto val : ans)
     {
         vec_ans.push_back(val);
@@ -88,41 +88,41 @@ vector<vector<int>> fourSum2(vector<int> &nums, int target)
 }
 
 // ----------------meth3
-vector<vector<int>> fourSum3(vector<int> &nums, int target)
+vector<vector<ll>> fourSum3(vector<ll> &nums, ll target)
 {
-    unordered_set<int> seen;
-    set<multiset<int>> ans_set;
-    int n = nums.size();
-    for (int i = 0; i < n; i++)
+    unordered_set<ll> seen;
+    set<multiset<ll>> ans_set;
+    ll n = nums.size();
+    for (ll i = 0; i < n; i++)
     {
-        for (int j = i + 1; j < n; j++)
+        for (ll j = i + 1; j < n; j++)
         {
-            for (int k = j + 1; k < n; k++)
+            for (ll k = j + 1; k < n; k++)
             {
-                int lastNumber = target - (nums[i] + nums[j] + nums[k]);
+                ll lastNumber = target - (nums[i] + nums[j] + nums[k]);
                 if (seen.count(lastNumber))
-                    ans_set.emplace(multiset<int>{nums[i], nums[j], nums[k], lastNumber});
+                    ans_set.emplace(multiset<ll>{nums[i], nums[j], nums[k], lastNumber});
             }
         }
         seen.insert(nums[i]);
     }
-    vector<vector<int>> ans(ans_set.size());
+    vector<vector<ll>> ans(ans_set.size());
     for_each(begin(ans_set), end(ans_set), [&, i(0)](auto &el) mutable
-             { ans[i++] = vector<int>(begin(el), end(el)); });
+             { ans[i++] = vector<ll>(begin(el), end(el)); });
     return ans;
 }
-// --------------meth4---->two pointer
-vector<vector<int>> fourSum4(vector<int> &nums, int target)
+// --------------meth4---->two poller
+vector<vector<ll>> fourSum4(vector<ll> &nums, ll target)
 {
     sort(nums.begin(), nums.end());
-    int n = nums.size();
-    vector<vector<int>> ans;
-    for (int i = 0; i < n; ++i)
+    ll n = nums.size();
+    vector<vector<ll>> ans;
+    for (ll i = 0; i < n; ++i)
     {
-        for (int j = i + 1; j < n; ++j)
+        for (ll j = i + 1; j < n; ++j)
         {
-            int l = j + 1, r = n - 1;
-            int remain = target - nums[i] - nums[j];
+            ll l = j + 1, r = n - 1;
+            ll remain = target - nums[i] - nums[j];
             while (l < r)
             {
                 if (nums[l] + nums[r] == remain)
@@ -153,9 +153,9 @@ vector<vector<int>> fourSum4(vector<int> &nums, int target)
 
 int main()
 {
-    vector<int> arr{1, 0, -1, 0, -2, 2};
-    int tar = 0;
-    vector<vector<int>> ans = fourSum1(arr, tar);
+    vector<ll> arr{1, 0, -1, 0, -2, 2};
+    ll tar = 0;
+    vector<vector<ll>> ans = fourSum1(arr, tar);
     for (auto val : ans)
     {
         for (auto v : val)
